Move all reports into components
